name and extension (of the selected object) 2
This can speed up the work of entering the name and extension for each shape. In this case, the name of the shape will be the same as the name of the object selected in the blender, that is, it is necessary to rename the object (inside the blender) with the same name as the exported shape.